Use color names in COLOR for bright variants
In the following text with pseudo-escapes:
[color bright-red] text [color green] here

The word "here" should be regular green and not bright green. By interpreting
the escape as bold, it doesn't get cleared the next time a color change
occurs. This causes the "here" in the above example to be bright green.

This commit changes it so the already defined bright color names in COLORS are used
instead, fixing the above issue.